ledger: define the tableledger interface
the tableledger module will implement the "model" for gorestable,
acting as data structure to hold all the reservation.

We pack the core functionality into a core module. While its API
is provisional, it is going to reside under internal/pkg, to be
eventually graduated in pkg once the API is stable.